switch to uint128

+++ ethereum/contracts/ERC20Vault.sol
@@ -15,23 +15,23 @@ contract ERC20Vault is Ownable {
     /// @param sender The address of the sender.
     /// @param token The address of the ERC20 token.
     /// @param amount The amount being deposited.
-    event Deposit(address account, address sender, address token, uint256 amount);
+    event Deposit(address account, address sender, address token, uint128 amount);
 
     /// @dev Emitted when funds are withdrawn.
     /// @param account The address of the ERC20App contract.
     /// @param recipient The address of the sender.
     /// @param token The address of the ERC20 token.
     /// @param amount The amount being withdrawn.
-    event Withdraw(address account, address recipient, address token, uint256 amount);
+    event Withdraw(address account, address recipient, address token, uint128 amount);
 
     /* State */
-    mapping(address => uint256) public balances;
+    mapping(address => uint128) public balances;
 
     /// @dev Accepts a ERC20 Token from the caller.
     /// @param _sender The address of the sender.
     /// @param _token The address of the Token.
     /// @param _amount The amount being deposited.
-    function deposit(address _sender, address _token, uint256 _amount) 
+    function deposit(address _sender, address _token, uint128 _amount) 
         external
         onlyOwner
     {
@@ -47,7 +47,7 @@ contract ERC20Vault is Ownable {
     /// @param _recipient The address that will receive funds.
     /// @param _token The address of the Token.
     /// @param _amount The amount being deposited.
-    function withdraw(address _recipient, address _token, uint256 _amount)
+    function withdraw(address _recipient, address _token, uint128 _amount)
         external
         onlyOwner
     {
